+++ src/components/policy-editor/ActionRowTimeout.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Row from 'react-bootstrap/Row';
-import { OptionalNumberField } from '../forms/OptionalNumberField';
+import { OptionalNumberField } from '../../forms/OptionalNumberField';
 import { LabelColumn } from './LabelColumn';
 import { WideValueColumn } from './WideValueColumn';
 import { EffectiveValue } from './EffectiveValue';
