+++ src/components/policy-editor/EffectiveBooleanValue.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import Form from 'react-bootstrap/Form';
-import { getDeepStateProperty } from '../deepstate';
+import { getDeepStateProperty } from '../../utils/deepstate';
 import { EffectiveValueColumn } from "./EffectiveValueColumn";
 
-export function EffectiveValue(component, policyField) {
+export function EffectiveBooleanValue(component, policyField) {
     const dsp = getDeepStateProperty(component, "resolved.definition." + policyField, undefined);
 
     return <EffectiveValueColumn>
         <Form.Group>
-            <Form.Control
+            <Form.Check
                 data-testid={'effective-' + policyField}
                 size="sm"
-                value={getDeepStateProperty(component, "resolved.effective." + policyField, undefined)}
+                checked={getDeepStateProperty(component, "resolved.effective." + policyField, undefined)}
                 readOnly={true} />
             <Form.Text data-testid={'definition-' + policyField}>
                 {component.PolicyDefinitionPoint(dsp)}
