+++ src/components/policy-editor/EffectiveTextAreaValue.jsx
@@ -1,17 +1,19 @@
 import React from 'react';
 import Form from 'react-bootstrap/Form';
-import { getDeepStateProperty } from '../deepstate';
+import { getDeepStateProperty } from '../../utils/deepstate';
 import { EffectiveValueColumn } from "./EffectiveValueColumn";
 
-export function EffectiveBooleanValue(component, policyField) {
+export function EffectiveTextAreaValue(component, policyField) {
     const dsp = getDeepStateProperty(component, "resolved.definition." + policyField, undefined);
 
     return <EffectiveValueColumn>
         <Form.Group>
-            <Form.Check
+            <Form.Control
                 data-testid={'effective-' + policyField}
                 size="sm"
-                checked={getDeepStateProperty(component, "resolved.effective." + policyField, undefined)}
+                as="textarea"
+                rows="5"
+                value={getDeepStateProperty(component, "resolved.effective." + policyField, undefined)}
                 readOnly={true} />
             <Form.Text data-testid={'definition-' + policyField}>
                 {component.PolicyDefinitionPoint(dsp)}
