+++ src/components/policy-editor/EffectiveValue.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Form from 'react-bootstrap/Form';
-import { getDeepStateProperty } from '../deepstate';
+import { getDeepStateProperty } from '../../utils/deepstate';
 import { EffectiveValueColumn } from "./EffectiveValueColumn";
 
-export function EffectiveListValue(component, policyField) {
+export function EffectiveValue(component, policyField) {
     const dsp = getDeepStateProperty(component, "resolved.definition." + policyField, undefined);
 
     return <EffectiveValueColumn>
@@ -11,8 +11,6 @@ export function EffectiveListValue(component, policyField) {
             <Form.Control
                 data-testid={'effective-' + policyField}
                 size="sm"
-                as="textarea"
-                rows="5"
                 value={getDeepStateProperty(component, "resolved.effective." + policyField, undefined)}
                 readOnly={true} />
             <Form.Text data-testid={'definition-' + policyField}>
