+++ src/components/policy-editor/PolicyEditor.jsx
@@ -7,14 +7,14 @@ import Form from 'react-bootstrap/Form';
 import Row from 'react-bootstrap/Row';
 import Spinner from 'react-bootstrap/Spinner';
 import Accordion from 'react-bootstrap/Accordion';
-import { handleChange, stateProperty, valueToNumber } from '../forms';
-import { StringList } from '../forms/StringList';
-import { LogDetailSelector } from '../forms/LogDetailSelector';
-import { OptionalBoolean } from '../forms/OptionalBoolean';
-import { OptionalNumberField } from '../forms/OptionalNumberField';
-import { RequiredBoolean } from '../forms/RequiredBoolean';
-import { TimesOfDayList } from '../forms/TimesOfDayList';
-import { errorAlert, PolicyEditorLink, sourceQueryStringParams, toAlgorithmOption } from '../uiutil';
+import { handleChange, stateProperty, valueToNumber } from '../../forms';
+import { StringList } from '../../forms/StringList';
+import { LogDetailSelector } from '../../forms/LogDetailSelector';
+import { OptionalBoolean } from '../../forms/OptionalBoolean';
+import { OptionalNumberField } from '../../forms/OptionalNumberField';
+import { RequiredBoolean } from '../../forms/RequiredBoolean';
+import { TimesOfDayList } from '../../forms/TimesOfDayList';
+import { errorAlert, PolicyEditorLink, sourceQueryStringParams, toAlgorithmOption } from '../../utils/uiutil';
 import { LabelColumn } from './LabelColumn';
 import { ValueColumn } from './ValueColumn';
 import { WideValueColumn } from './WideValueColumn';
