+++ src/components/policy-editor/SectionHeaderRow.jsx
