+++ src/components/policy-editor/WideValueColumn.jsx
