--- src/setupProxy.js
@@ -1,11 +1,0 @@
-const { createProxyMiddleware } = require('http-proxy-middleware');
-
-module.exports = function(app) {
-  app.use(
-    '/api',
-    createProxyMiddleware({
-      target: 'http://127.0.0.1:51515',
-      changeOrigin: true,
-    })
-  );
-};
